integration-examples/apiScripts/requirements.txt
jsonpath_ng
pyyaml
requests